// src/settings/enums.ts
export enum ADDRESS_TOKENS_BSC_ENUM {
  SKY = "0x0203a4ce70cC5E0566e1122f071B59d2ed982D67",
  PIPI = "0xf86E639Ff387b6064607201A7a98F2c2B2FEB05f",
  TOTO = "0x508EbA399d39B84e46c2E78cB0580c06A1930E1B",
  BABYPIPI = "0x5FC52BDDb1AB30C2ef733F7881830Fa04f07cb09",
}

export enum ADDRESS_TOKENS_ETH_ENUM {
  PIPI = "0x18D84f4305b24DF99c58d7Ba92D31213d68b5455",
}

export enum ADDRESS_TOKENS_POLYGON_ENUM {
  PIPI = "0x250606C2870147d8cf1Ab242466A38bA6646B8B8",
}

export enum ADDRESS_TOKENS_MUMBAI_ENUM {
  PIPI = "0x823940ef02380CC2c614e00CC6218421C03a159D",
  SIMPLE = "0x214aae7bdd887fd5640599727ca2851830d185b5",
}

export enum ADDRESS_TOKENS_CORE_ENUM {
  PIPI = "0x3034802fc4C9A278D0886eD77fd3F79fd789c898",
}
export enum ADDRESS_TOKENS_CORE_TESTNET_ENUM {
  CAT = "0x9ef761219423a06752f8552c6033ab7b07bd6122",
}
export enum ADDRESS_TOKENS_GENESYS_ENUM {
  PIPI = "0x250606C2870147d8cf1Ab242466A38bA6646B8B8",
}
export enum ADDRESS_TOKENS_BASE_ENUM {
  PIPI = "0xf0910396FCEF3b4316CfcaebEF22A44Ad0D99584",
}

export enum CRYPTO_ENUM {
  BSC = 56,
  ETH = 1,
  POLYGON = 137,
  MUMBAI = 80001,
  CORE = 1116,
  CORE_TESTNET = 1115,
  GENESYS = 16507,
  BASE = 8453,
  SONEIUM = 1868
}